server/pyproject.toml
[project]
name = "espdisplayserver"
version = "0.1.0"
description = "A Server for ESPDisplay"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic>=2.11.7",
    "python-dotenv",
    "pyyaml"
]

[tool.ruff]
lint.ignore = [
    "E402",  # allow code before imports
    "F401",  # unused imports
    "F841",  # unused variables
]